[Rename] StoryCardEntity temp2 -> common
