--- src/main/java/com/vlc/maeummal/domain/template/common/TemplateRepository.java
@@ -1,19 +1,0 @@
-//package com.vlc.maeummal.domain.template.common;
-//
-//import org.springframework.data.jpa.repository.Query;
-//import org.springframework.stereotype.Repository;
-//
-//
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//import java.util.List;
-//
-//@Repository
-//public interface TemplateRepository extends JpaRepository<TemplateEntity, Long> {
-//    // JpaRepository에서 제공하는 기본 CRUD 메서드와 추가적인 쿼리 메서드를 정의할 수 있습니다.
-//    @Query(
-//
-//    )
-//    List<T> findAll();
-//}
-//
