+++ src/main/java/com/vlc/maeummal/domain/template/template2/service/Template2Service.java
@@ -3,7 +3,7 @@ package com.vlc.maeummal.domain.template.template2.service;
 
 import com.vlc.maeummal.domain.template.template2.dto.Template2RequestDTO;
 import com.vlc.maeummal.domain.template.template2.dto.Template2ResponseDTO;
-import com.vlc.maeummal.domain.template.template2.entity.StoryCardEntity;
+import com.vlc.maeummal.domain.template.common.StoryCardEntity;
 import com.vlc.maeummal.domain.template.template2.entity.Template2Entity;
 import com.vlc.maeummal.domain.template.template2.repository.StoryCardRepository;
 import com.vlc.maeummal.domain.template.template2.repository.Template2Repository;
